% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funique.R
\name{funique}
\alias{funique}
\title{A faster unique() function}
\usage{
funique(x)
}
\arguments{
\item{x}{Input data. If x is not a data frame or date-time object, then it is
simply passed to \code{\link[base]{unique}}}
}
\value{
The unique values of x.
}
\description{
Similar to \code{\link[base]{unique}}, only optimized for working with
date-time and some list columns.
}
\examples{

## create data set with date-time columns
d <- datasets::mtcars
d$dttm <- Sys.time() + runif(nrow(d), -1000, 1000)

## create multiple data frames with duplicate rows
d <- lapply(1:50, function(.) rbind(d, d[sample(seq_len(nrow(d)), 20), ]))

## merge into single data frame
d <- do.call("rbind", d)

## speed test
library(microbenchmark)
(mn <- microbenchmark(unique(d), funique(d)))

## check output
identical(unique(d), funique(d))

}
